AI Agents/Turjuman_OCR/requirements.txt
fastapi
uvicorn
python-multipart
pillow
google-generativeai
python-dotenv
openai